Update map() call to work with Python 2 or Python 3.
The map() call that applies the crop function to the training data
is automatically applied on the fly in Python 2,
but returns a generator in Python 3.
This creates problems when converting the resulting function map
to a numpy array - the numpy array is empty in Python 3.
This fixes the problem by explicitly converting the generator
to a list. This is about as fast as pre-allocating an empty numpy array
and populating it in a for loop. See https://gist.github.com/charlesreid1/436c51eae1e6a0d011d86f3796dec853